LayoutTests/fast/workers/worker-messageport-expected.txt

CONSOLE MESSAGE: line 17: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that pages and workers can send MessagePorts to one another. Should print "DONE" when done.
